Refactor DealDashboard and LeadPage to streamline status handling and UI components; introduce StatusMenu and DealItem for improved modularity and maintainability.
